backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
openai
groq
psycopg2-binary